Implemented UserService and implementation with tests validating various usecases

+++ services/src/main/java/nl/shanelab/kwetter/services/impl/UserServiceImpl.java
@@ -1,50 +1,107 @@
 package nl.shanelab.kwetter.services.impl;
 
+import lombok.NoArgsConstructor;
 import nl.shanelab.kwetter.dal.dao.UserDao;
 import nl.shanelab.kwetter.dal.domain.Role;
 import nl.shanelab.kwetter.dal.domain.User;
 import nl.shanelab.kwetter.dal.qualifiers.InMemoryDao;
 import nl.shanelab.kwetter.services.UserService;
+import nl.shanelab.kwetter.services.exceptions.UserAlreadyExistsException;
+import nl.shanelab.kwetter.services.exceptions.UserException;
+import nl.shanelab.kwetter.services.exceptions.UserIncorrectCredentialsException;
+import nl.shanelab.kwetter.services.exceptions.UserNotFoundException;
 
 import javax.ejb.Stateless;
 import javax.inject.Inject;
 import java.util.Collection;
 
 @Stateless
+@NoArgsConstructor
 public class UserServiceImpl implements UserService {
-    
-    @InMemoryDao
+
     @Inject
+    @InMemoryDao
     UserDao userDao;
 
-    public User register(String name, String password) {
+    public User register(String name, String password) throws UserException {
         return this.register(name, password, Role.USER);
     }
 
-    public User register(String name, String password, Role role) {
-//        User user = this.getByUserName(name); throw (custom?) error when username already exist
+    public User register(String name, String password, Role role) throws UserException {
+        if (name == null || password == null) {
+            throw new IllegalArgumentException();
+        }
+        if (this.getByUserName(name) != null) {
+            throw new UserAlreadyExistsException(name);
+        }
+
         return userDao.create(new User(name, password, role));
     }
 
-    public User signIn(String name, String password) {
-        User user = this.getByUserName(name); // throw (custom?) error when no such user exists
+    public User signIn(String name, String password) throws UserException {
+        if (name == null || password == null) {
+            throw new IllegalArgumentException();
+        }
 
-        return user != null && user.getPassword().equals(password) ? user : null;
+        User user = this.getByUserName(name);
+        if (user == null) {
+            throw new UserNotFoundException(name);
+        }
+        if (!user.getPassword().equals(password)) {
+            throw new UserIncorrectCredentialsException(name);
+        }
+
+        return user;
     }
 
-    public User edit(User user) {
+    public User rename(String name, User user) throws UserException {
+        if (name == null || user == null) {
+            throw new IllegalArgumentException();
+        }
+        if (this.getByUserName(name) != null) {
+            throw new UserAlreadyExistsException(name);
+        }
+
+        user.setUsername(name);
+
         return userDao.edit(user);
     }
 
-    public void remove(User user) {
+    public User setBiography(String bio, User user) throws UserException {
+        if (bio == null || user == null) {
+            throw new IllegalArgumentException();
+        }
+        if (this.getById(user.getId()) == null) {
+            throw new UserNotFoundException(user.getId());
+        }
+
+        return userDao.edit(user);
+    }
+
+    public void remove(User user) throws UserException {
+        if (user == null) {
+            throw new IllegalArgumentException();
+        }
+        if (this.getById(user.getId()) == null) {
+            throw new UserNotFoundException(user.getId());
+        }
+
         userDao.remove(user);
     }
 
     public User getById(long id) {
+        if (id < 0) {
+            throw new IllegalArgumentException();
+        }
+
         return userDao.find(id);
     }
 
     public User getByUserName(String name) {
+        if (name == null) {
+            throw new IllegalArgumentException();
+        }
+
         return userDao.getByUsername(name);
     }
